fix(types): types moved to a single file
