--- src/afip-context.ts
@@ -1,50 +1,0 @@
-import { WSAuthTokens } from "./auth/types";
-
-export type Context = Omit<AfipContext, "ticketPath"> & { ticketPath?: string };
-
-export type AfipContext = {
-  /**
-   * Flag for production or testing environment
-   *
-   * @var boolean
-   **/
-  production?: boolean;
-
-  /**
-   * Content file for the X.509 certificate in PEM format
-   *
-   * @var string
-   **/
-  cert: string;
-
-  /**
-   * Content file for the private key correspoding to CERT (PEM)
-   *
-   * @var string
-   **/
-  key: string;
-
-  /**
-   * The CUIT to use
-   *
-   * @var int
-   **/
-  cuit: number;
-
-  /**
-   * Tokens object if you have one created before
-   *
-   * @var authTokens
-   **/
-  authTokens?: WSAuthTokens;
-
-  /**
-   * Flag that if is true, the access tickets data is handled by the developer, otherwise is saved locally.
-   */
-  handleTicket?: boolean;
-
-  /**
-   * The path of the auth obj if the package is auto managed
-   */
-  ticketPath: string;
-};
