+++ src/index.ts
@@ -1,5 +1,2 @@
 export * from "./afip";
-export * from "./interfaces";
-export * from "./afip-context";
 export * from "./auth/access-ticket";
-export * from "./auth/types";
