+++ src/soap/interfaces/PersonaServiceA5/PersonaServiceA5Port.ts
@@ -1,5 +1,5 @@
 import { Client } from "soap";
-import { SoapAsyncFunc } from "../soap-async-func";
+import { SoapAsyncFunc } from "../../../types";
 /* tslint:disable:max-line-length no-empty-interface */
 export interface IgetPersonaInput {
   /** http://a5.soap.ws.server.puc.sr/#xs:string(undefined) */
