--- src/soap/interfaces/soap-async-func.ts
@@ -1,5 +1,0 @@
-export type SoapAsyncFunc<I, O> = (
-  input: I,
-  options?: any,
-  extraHeaders?: any
-) => Promise<[O, string, { [k: string]: any }, string]>;
